refactor: rename to pascal case
