+++ apps/antalmanac/src/components/Calendar/CalendarCourseEvent.tsx
