+++ apps/antalmanac/src/components/Calendar/toolbar/ScheduleSelect/schedule-select-buttons/DeleteScheduleButton.tsx
